+++ tests/test_particles.py
@@ -24,8 +24,8 @@ def test_variable_init(fieldset, mode, npart=10):
         p_double = Variable('p_double', dtype=np.float64, initial=11.)
         p_int = Variable('p_int', dtype=np.int32, initial=12.)
     pset = ParticleSet(fieldset, pclass=TestParticle,
-                       lon=np.linspace(0, 1, npart, dtype=np.float32),
-                       lat=np.linspace(1, 0, npart, dtype=np.float32))
+                       lon=np.linspace(0, 1, npart),
+                       lat=np.linspace(1, 0, npart))
 
     def addOne(particle, fieldset, time):
         particle.p_float += 1.
@@ -68,21 +68,21 @@ def test_variable_special_names(fieldset, mode):
 def test_variable_init_relative(fieldset, mode, npart=10):
     """Test that checks relative initialisation of custom variables"""
     class TestParticle(ptype[mode]):
-        p_base = Variable('p_base', dtype=np.float32, initial=10.)
-        p_relative = Variable('p_relative', dtype=np.float32,
+        p_base = Variable('p_base', dtype=np.float64, initial=10.)
+        p_relative = Variable('p_relative', dtype=np.float64,
                               initial=attrgetter('p_base'))
-        p_offset = Variable('p_offset', dtype=np.float32,
+        p_offset = Variable('p_offset', dtype=np.float64,
                             initial=attrgetter('p_base'))
-        p_lon = Variable('p_lon', dtype=np.float32,
+        p_lon = Variable('p_lon', dtype=np.float64,
                          initial=attrgetter('lon'))
-        p_lat = Variable('p_lat', dtype=np.float32,
+        p_lat = Variable('p_lat', dtype=np.float64,
                          initial=attrgetter('lat'))
 
         def __init__(self, *args, **kwargs):
             super(TestParticle, self).__init__(*args, **kwargs)
             self.p_offset += 2.
-    lon = np.linspace(0, 1, npart, dtype=np.float32)
-    lat = np.linspace(1, 0, npart, dtype=np.float32)
+    lon = np.linspace(0, 1, npart)
+    lat = np.linspace(1, 0, npart)
     pset = ParticleSet(fieldset, pclass=TestParticle, lon=lon, lat=lat)
     # Adjust base variable to test for aliasing effects
     for p in pset:
